requirements.txt
runpod==1.7.0
httpx==0.27.2
docling>=2.5.0
langchain-docling>=2.0.0
langchain-core>=0.3.0
langchain-ollama>=0.2.0
langchain-huggingface
langchain-text-splitters
python-dotenv
fastapi
uvicorn
onnxruntime-gpu
easyocr
psycopg2-binary
pgvector
streamlit
pandas
numpy
plotly
requests